Hack, make external trigger work with guppy
This is a hack commit to make the external trigger work with
the GUPPY camera. This camera doesn't report any trigger input,
which was not expected. For now the external trigger is now
activated, as soon as the  FrameStartTriggerEventToEdgeRising
attribute is set, which is a hack. Also the availability of
this attribute is hardcoded